Review comments, move logic to abstract method
