+++ app/src/main/java/io/homeassistant/companion/android/qs/Tile2Service.kt
@@ -1,39 +1,21 @@
 package io.homeassistant.companion.android.qs
 
 import android.os.Build
-import android.service.quicksettings.TileService
+import android.service.quicksettings.Tile
 import androidx.annotation.RequiresApi
-import io.homeassistant.companion.android.common.dagger.GraphComponentAccessor
-import io.homeassistant.companion.android.common.data.integration.IntegrationRepository
-import javax.inject.Inject
 
 @RequiresApi(Build.VERSION_CODES.N)
-class Tile2Service : TileService() {
+class Tile2Service : TileExtensions() {
 
-    @Inject
-    lateinit var integrationUseCase: IntegrationRepository
+    override fun getTile(): Tile {
+        return qsTile
+    }
+
+    override fun getTileId(): String {
+        return TILE_ID
+    }
 
     companion object {
         private const val TILE_ID = "tile_2"
-    }
-
-    override fun onClick() {
-        super.onClick()
-        DaggerTileComponent.builder()
-            .appComponent((applicationContext as GraphComponentAccessor).appComponent)
-            .build()
-            .inject(this)
-        TileExtensions.setTileData(applicationContext, TILE_ID, qsTile)
-        TileExtensions.tileClicked(applicationContext, TILE_ID, qsTile, integrationUseCase)
-    }
-
-    override fun onTileAdded() {
-        super.onTileAdded()
-        TileExtensions.setTileData(applicationContext, TILE_ID, qsTile)
-    }
-
-    override fun onStartListening() {
-        super.onStartListening()
-        TileExtensions.setTileData(applicationContext, TILE_ID, qsTile)
     }
 }
